Rename search-page folder to search
