+++ app/components/search/SearchResult.js
